Adding support to manifests with segmented timeline (remaning bug where it fails to download last segment)

+++ src/main/java/net/ericsson/emovs/download/models/Manifest.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
  */
 
 public class Manifest {
-    public Long durationSeconds;
+    public double durationSeconds;
     public ArrayList<AdaptationSet> adaptationSets;
 
     public Manifest() {
